Support specifying an output file for the fetch command.
This will hopefully make the repository history much less noisy, as it allows for re-downloading a file without a separate rename step.

The following corner cases have been considered:

- The output file is passed through realpath(join(hackenv, output)) to normalize out any symbolic links.

- If hackenv (with an enforced trailing /) isn't a prefix of the normalized output file, the action is refused.

- If the target output file would be a directory, the action is also refused.

As an aside, also add a `--` argument in front of the URL to stop wget from interpreting it as an argument.