Add -- to separate command from arguments for umlbox.
The newly refactored umlbox wrapper script uses argparse, which
doesn't have the previous behavior of treating all arguments from the
first non-flag one as positional. This commit adds a -- in the command
line to make sure remaining arguments are treated as part of the
command to run in the sandbox.

It also uses Python 3 and subprocess.run, since Python 2 is eventually
going away.